requirements.txt
requests>=2.31.0
rich>=13.7.0
googlesearch-python>=1.2.3
fake-useragent>=1.4.0
google-search-results>=2.4.2
